Disable translatedRotatedAndScaledPainter testcase as it is broken.
The problem is that the order of rounding to fixedpoint is different
in the static and normal text case and this leads to some glyphs
occationally being 1 pixel off.

In the normal case, the glyphs are laid out and positioned in fixed
point, then transformed using the full current matrix including the
position which is then snapped to integer device space when drawn.

In the static text case, the glyphs are laid out and positioned in
fixed point related to 0,0. The actual device position is then
added to the glyph positions.