QFileDialog::HideNameFilterDetails breaks Cocoa QFileDIalog filter
Using namefilters in the non-native file dialog is a bit broken by
design, since it don't store the whole selected name filter, but
only the filter with details stripped off (if this is specified). And
it stores this data in gui widgets. And the native file dialog also
suffers from this. At least this patch will pick the first matching
name filter (with details) given a details-stripped version.

Task-number: QTBUG-12870
Reviewed-by: cduclos