Revised fix for pixmap loading
Change 2c8f9dee611 had an unwelcome side-effect and was reverted.

Since we do have the undocumented behavior that we scan for all possible
extensions when QPixmap is provided with a filename without extensions
we could not simply exit if the file did not exist. We now check if
there are extensions before doing so.

Reviewed-by: gabi
Task-number: QTBUG-11137