Fixed touch event delivery in QGraphicsView.
When a touch event with a second touch pressed is delivered inside
graphicsview, we should combine it with the closest touch point even if the
item under the second touch is not direct ancestor or child of the first
touches' target item.

So adding a second touch inside the item's bounding rect will send a
TouchUpdate event to the item instead or starting a new touch event sequence.

Task-number: QT-3795
Reviewed-by: Bradley T. Hughes