Call eglReleaseThread() when a thread exits, and fix warnings.
In Qt we track the current context in a thread via a QThreadStorage.
The thread storage contents are deleted in the thread context, just
before it's destroyed. This means we can safely call eglReleaseThread()
in the QGLThreadContext destructor. We can then get rid of unnecessary
context swaps, since we don't need to take care to reset a thread
context back to 0 after having used e.g. the QGLShareContextScope
mechanism, which is good.

Reviewed-by: Samuel