Fix mispositioned text with QStaticText and OpenVG graphics system
The OpenVG paint engine, like the OpenGL2 paint engine, supports
caching the untransformed glyphs and transforming them as they are
drawn. Since we would pretransform the positions of the glyphs, the
transformation would be applied twice, thus making the glyphs appear
in the wrong location when the painter had a transform set.

Task-number: QTBUG-13049
Reviewed-by: Gunnar