Removed QDirPrivate layer of indirection
Merged QDirPrivate with QDirPrivate::Data, as QDirPrivate served no
purpose by itself, only adding an additional layer of indirection to the
potentially shared private data, and an unnecessary allocation.

Now, QDir holds a QSharedDataPointer to its private data. Private data
will be shared among copied instances with COW semantics. Still, this
sharing is very limited as plenty of regular use cases will cause the
shared data to detach, such as refreshing the file lists.

As the use QSharedDataPointer breaks usage of the Q_DECLARE_PRIVATE
macro, we manually define the d_func'tions. Non-const d_func detaches on
shared data. A detach function was added to the public interface to
support this. (On a side note, QFileInfo already exposes a similar
detach function).

As much as possible, detach is handled implicitly inside the Q_D macro,
through the d_func() non-const overload.  On the other hand, implicit
creation of file engines through detach was made explicit with a call to
a new initFileEngine function.

Reviewed-by: Thomas Zander