Don't rely on mapFromGlobal in QGraphicsScenePrivate::itemsAtPosition.
The old approach discovered a bug in QWidget::mapToGlobal in the context of
lighthouse, which in turn brought our eyes to
QGraphicsScenePrivate::itemsAtPosition. The implementation itself was
correct, but we can easily remove the-not-so-cheap mapToGlobal call
since we already have the scene position, which is all we need.

This functionality is heavily tested by existing auto tests, and all of
them pass.

Reviewed-by: mae