Fix cache_cost initialization of QFontEngineMultiWin
Originally, QFontEngineMultiWin does not set its cache_cost according to
the regular font engines it refers to, which makes cache management in
QFontCache useless because cache_cost is always 0, and garbage
collecting timerEvent can hardly be triggered. This patch solve part of
the problem by using cache_cost of the first font engine for
QFontEngineMultiWin.

However, if a font engine is loaded by QFontEngineMulti::loadEngine, the
cost is still not counted.

Task-number: QTBUG-12562
Reviewed-by: Eskil Abrahamsen Blomfeldt