Cocoa: Demo browser can get stuck after closing modal dialog
This turns out to be a glitch in the event dispatcher for Qt/Cocoa.
The problem is that you can, in some really slim cases it must be, end up
calling processEvents with the exec flag set, while the current modal
dialog has been told to leave modal. In that case, we recurse into
a new event dispatching loop for a session that is about to die. And
that causes problems when the session actually dies. This patch makes
sure that you cannon start spinning a session that is about to die, by
setting the sessionCache to zero.

Task-number: QTBUG-13164
Reviewed-by: Tor Arne
(cherry picked from commit a373648fa876a9684394aa9b7ecafd51e53f4742)