Fixed hiding a window on X11 before it was shown.
On X11 show() just issues a request to show a window, so if someone tries to
call hide() before the window was successfully mapped to the screen by the
window manager it will not work (on most window managers unfortunately).

So instead, when hide() is called we internally mark the window as "no need to
show anymore" and then if we still get MapNotify event, we will hide the window
right away.

Reviewed-by: Markus Goetz
Reviewed-by: Bradley T. Hughes