Make sure propagated font in QGraphicsWidget can be set on a QPainter
In QWidget and QGraphicsWidget, propagating fonts blank out the resolve
mask, mainly to help in determining inheritance. This means that when
the font is set on a QPainter with device and therefore resolved against
the device font, none of the attributes in the font will take presedence
and the font will essentially just be ignored. For QWidgets this has not
been a problem due to the fact that the painter is opened on the widget,
getting the propagated font as its device font. For QGraphicsWidget,
however, the QPainter is opened on the viewport and the viewport's font
will therefore override the propagated font when it's passed into
setFont(). To fix this, we need the font returned by
QGraphicsWidget::font() to contain the resolve mask of the inherited
parts of its font as well. This is done last minute to avoid messing
up the rules of inheritance in the QGraphicsWidget tree (problem arises
e.g. in hierarchy A -> B -> C and a font is set on B which should override
the font C has previously inherited from A.)

Task-number: QTBUG-13188
Reviewed-by: Lars