QLineEdit cursor is shown when asked to be non-visible
If QLineEdit is sent a QInputMethodEvent::Cursor event with value zero,
widget still shows blinking cursor. This is due to that QLineControl
ignores its internal state variable m_hideCursor value.

As a fix, m_hideCursor is now considered when drawing the cursor.
Additionally, the cursor blinking timer is stopped when cursor is
hidden.

Task-number: QT-3670
Reviewed-by: axis