Reverting QDir::detach introduced earlier
From the client API it is a broken concept, because it could leave the
user with an uninitialized file engine. Fixing to initialize the file
engine, would mean it is useless for internal use where file engines are
initialized separately.

In the end, removing both the QDir::detach introduced earlier and
throwing away the private d_func'tions altogether is a binary-compatible
change that should allow us to push this change in a patch release
(fingers crossed!)

Reviewed-by: Thiago Macieira