Fix QtScript Date <--> QDateTime (local time) conversion
This has already been fixed in 4.7 (QTBUG-9770), but the change is
too big to backport.
The general idea is the same: Only operate on UTC dates internally
(since that's how JS dates are stored), and let QDateTime take care
of converting from/to local dates as necessary.

The fix itself shouldn't be merged to 4.7, but the autotests should.

Task-number: QTBUG-9770
Reviewed-by: Jedrzej Nowacki