fix windows build with -prefix
passing just a spec name to qmake is bogus, as it will try to find it in
the install dir, which of course won't work before it is installed. so
either pass nothing at all (when preparing the libraries, as
.qmake.cache already contains the right path), or explicitly pass a full
path (when building the host tools, as we have to override the spec
here).

Reviewed-by: mariusSO