Cleaned up position tracking in the Symbian input methods.
This fixed a case where using password mode would lead to Qt
believing that the cursor was one step to the left of where it really
was. This would have the effect of replacing the last character
instead of appending to it, and even crashing if the cursor was all
the way to the left.

The code is also much cleaner this way, but it meant that QTBUG-9867
had to be solved differently this time. We do this by assuming that
and empty FEP update means "erase last character", which seems to
work well in practice.

Also added a long overdue autotest for the FEP input methods. Most
tests pass, but some don't, which I will try to fix later.

Task:     QTBUG-9867
Task:     QTBUG-12949
RevBy:    Miikka Heikkinen
AutoTest: Included