Updated Twitter demo in response to the new Twitter oauth requirement
Twitter is blocking access to the old authentification method that
our demo was using. This breaks the demo completely. To fix this, I
changed the demo into only make use of the public search API. You
can no longer tweet but you can search for different tweets.

At the moment making an oauth capable demo is not desirable due to
the problems with handling a secret app token among other issues.

Task-number: QTBUG-13181
Reviewed-by: Alan Alpert