QS60Style: Itemviews are drawn incorrectly
When running QS60Style on hardware, all the itemview items without
any special background (i.e. not "pressed", not "highlighted", not
"alternate") are drawn with "pressed button" graphics.

This is due that the internal drawing function gets called without
any enum value set and it seem to pick the second enumeration constant
from theme element list. The enumeration constant is not defined
to match to any integer value.

As a solution, style will not call the drawing function when theme
element is not defined.

Task-number: QTBUG-11601
Reviewed-by: Liang Qi