Fixed a case where a newly created native widget would lose focus.
If a native control was created for a child widget, it would receive
the FocusChanged signal in its QSymbianControl with a value of not-
focused. This would lead the code to deactivate the whole toplevel
window. This was fixed by having Qt traverse the widget chain up to
the toplevel and check if parent widgets are also not focused.

AutoTest: Included and passed
RevBy:    Denis Dzyubenko
RevBy:    mread