Breaking changes to QFileSystemEntry's constructors
Since we are using a QString for the native file path on windows, the
constructors we had on other platforms didn't make sense: we can't have
two constructors taking a single QString parameter, and it would be
error prone to have the two argument constructor (taking two QString
parameters, on Windows) change the meaning of the first argument.

This patch introduces a typedef for the NativePath type being used; the
constructor taking the native path now requires a dummy parameter that
makes the intent expliciti; and the order of the arguments on the
constructor taking both native and Qt paths had the order switched.

Done-with: João Abecasis