Fixed input context trying to squeeze content into a full widget.
Problem was reproduced on N97. If the FEP detects that the widget is
full while still editing text, it will try to send those events as
key events instead. Since this screws up the content in the widget,
we stop those events from reaching the widget in the input context.

AutoTest: Passed
Task:     QTBUG-12949
RevBy:    Miikka Heikkinen