Discard empty buffer on call to QAudioOutput::resume()
The Symbian backend holds buffers which are passed to it from DevSound,
exposing the memory of these buffers via the QIODevice interface.

When QAudioOutput::resume() is called, the backend re-starts data flow
by passing the buffer it holds back to DevSound.  Previously, this
would not happen if the buffer was empty, potentially causing playback
to stall.  This patch ensures that the buffer is always sent back to
DevSound, and data flow therefore always resumes.

Task-number: QTBUG-13059
Reviewed-by: Derick Hawcroft