Ensure that OpenGL contexts are attached to an NSView before first paint
Attaching an OpenGL context to an NSView too soon will result in an
error "invalid drawable" printed on the console. We used to guard
against this by checking the visible property of the NSWindow, but this
turned out to be too late, as we had already recived an initial paint
event by then as part of showing the window. The visual result was a
single frame of gray painted before the user's paint event code took
effect.

We solve this by hooking into setInitialFirstResponder on the NSWindow,
which is called as part of making the window visible for the first time.
At this point it's safe to attach the GL context to the NSView, so we
iterate all the GLWidget children of the top level window and make sure
the context is attached by sending a MacGLWindowChange event.

The check in qt_mac_update_child_gl_widgets() for a top level window
had to be removed for this approach to work, but should be okey as
we're only iterating the children.

Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@nokia.com>
Reviewed-by: Trond Kjernåsen <trond.kjernasen@nokia.com>