Use the QSharedData and remove dependency on the QObjectPrivate.
Current implementation strongly relay on a reference counting, all
private classes inherits from the QSharedData and their instances
are removed only if they are not used anymore. It simplify the
code as we don't need to check if pointers are still valid.

I removed dependency on the QObjectPrivate as it is not used directly
by the QScriptEngine. It is a private class and if we want to
modularize Qt then we shouldn't use private classes. But the most
important is that we need QSharedDataPointer as q_ptr not
QScopedPointer, which would delete our internal objects too quickly.