Fix over-stating in QAbstractFileEngine::create
With recursive entry resolution, we might end up querying the engine for
an entry once per recursion level. There was also the possibility that
stale data would remain in the meta data instance.

When resolving an entry, we now check for its existence at the leaf
level and handle clean-up right away.

Reviewed-by: Thomas Zander