Replaced QWeakPointer to *Pointer in QGestureManager.
The QWeakPointer was part of a key in a QMap. When the QWeakPointer gets deleted,
the key changes and the sorting of the Map was wrong. That leads to an random crash.

It should be safe not to use QWeakPointer as we remove pointers from the
objectGestures map when a widget or graphicsitem gets destroyed.

Reviewed-by: Denis Dzyubenko