QDataStream doesn't check return values from QIODevice::write
QDataStream does not check values from QIODevice::write and QIODevice::read
since it cannot decide what to do in case of errors or less data written
or read. The fix for this is to fix the documentation for
QIODevice::writeData and QIODevice::readData to make sure implementations
write and read all the requested data.

Task-number: QTBUG-13108
Reviewed-by: David Boddie