Update the error handling of invalid hostnames in QUrl.
This should let QUrl::errorString() show the original, invalid hostname
in the URL it reports to be invalid.

There may be a side-effect that QUrl::toEncoded() may include the broken
hostname on the first call, then miss it in the next calls. But since
QUrl::isValid() is returning false, we can consider that the result of
toEncoded() is undefined, so anything goes.

Reviewed-by: Olivier Goffart