Fix text size issues for high ppi displays
Many layout sizes in flickr are specified by the artwork sizes.
The heights and widths are defined in static pixel sizes.
Therefore the text has to obey the size restrictions, and should
not overlap.

This patch sets the font sizes in pixels. Otherwise the default
font sizes in points may be too big on high ppi displays.
(XM5800 has ~230 ppi)