Make QFileInfo use the new filesystemengine
The creation of a QAbstractFileEngine derived class will now
be avoided if the file we are watching is on the native file system
and from then on all access will be through the direct API

Notice that for QFileInfo::path, isRelative, filePath, fileName,
baseName, completeBaseName, completeSuffix and suffix we now use
the QFileSystemEntry instead of asking the engine.  This means
that some buggy or just weird user provided engines no longer
get used and as such we might not be bug compatible.

Reviewed-by: João Abecasis