Editable QComboBox popup immediately closes itself after a click
This is due that timer that blocks for while mouse release events
(blockMouseReleaseTimer) is not started. There is an optimization
in QComboBox that tries to avoid calling d->viewContainer() since
that method creates the viewContainter if it doesn't exist.
Unfortunately this optimization does not take into account if the
container is already created.

In this particular case, viewContainer is already created, but
since QComboBox creates a QLineEdit, it skips the timer start,
which leads to situation where combobox's popup menu receiving
mouse release event and thus closes itself.

As a fix, blockMouseReleaseTimer is started if the viewContainer is
already created.

Task-number: QTBUG-13231
Reviewed-by: axis