Fixed crash in input methods when using symbols menu and numbers only
When inputting only numbers, the symbol menu should not do anything.
However in the old code the resource id of the symbol table was still
being set, so the symbol key on N97 would look up a table that was
not valid for the current input mode and crash.

Fixed by setting the symbol table id to zero under those conditions.

RevBy:    Sami Merila
Task:     QTBUG-13472
AutoTest: Included