Cleanup. Make QSEP::m_metaObjectTemplate private.
I believe that some aspects of implementation should be really private.
The QSVP doesn't need information about m_metaObjectTempalate, it doesn't
need to know how meta stuff is used by engine, all it needs is to check
if a given value is a QMetaObject instance. The QSEP can and should
provide this information.

This design is more consistent as we have isQtVariant and isQtSignal
that do a similar thing.