New stable autotests were added.
In general the new testsuite is a subset of standard one, only in rare
situation it adds something. The main advantage of it is that it is
passing now (OK, mostly QSV have 18 failures and QSE 3).

Ideally new test should be merged with standard ones. But for now it is
nice to have something that checks basic functionality and gives a few
failures instead of thousands. Lets keep them passing.