Fix QFileInfo::isHidden on Mac
Having HiddenAttribute as part of PosixStatFlags, meant that we would
assume we knew about it whenever we stat'ed, which wasn't the intent. As
that is just one of the queries we do on Mac OS to determine if a file
is considered hidden.

Instead, we explicitly ask for a stat anytime the attribute is queried
for.

The proper fix will be to perform the cheap operations (e.g., file name
begins with '.') first and only stat if we still haven't got an answer.