QScriptEngine: Some handle/context scope fixes.
The QScriptEngine use the QV8Context now, it is nicer to use
the same approach (class) everywhere.

This patch is preparation for setGlobalObject implementation. There is
a subtle difference between calling QV8Context and v8::Context::Scope
although they have the same functionality. The first one is delegating
change of a context to an engine. It means that QSEP::m_context can be
easly changed at any point by the engine.

The patch fix undefinedValue test.