Remove redundant code in SgImage -> VGImage conversion on Symbian.
Calling the various Khronos getError() functions here is inefficient
since it involves a round trip to the GPU and gains nothing since the
null handle can be used to test for error conditions at no cost. Also
no need to open the RSgDriver because we don't actually need to call
any functions on RSgImage, just convert it.

Reviewed-by: Alessandro Portale
Reviewed-by: Jani Hautakangas