This patch allows modifications on PropertyChanges on the fly
This is important for Bauhaus to ensure that we can
edit properties and states dynamically in the visual
editor

Most of the changes consist of added functions
for Bauhaus, that do not affect any current
behaviour.

Important changes:

QDeclarativeAbstractBinding is holding a
QSharedPointer to itself so that other classes
can hold a weak reference to avoid accessing
dangling pointers while keeping track of bindings.

QDeclarativeStateOperation now has its own private class
QDeclarativeStateOperationPrivate to hold a pointer back
to QDeclarativeState since the parent is not set correctly
in every context.

QDeclarativePropertyChangesPrivate does now hold a
QDeclarativeGuard instead of a raw pointer to the
target QObject. In Bauhaus the target object might
be deleted.

Reviewed-by: Aaron Kennedy