fix error messages when configuring Qt
Commit 019032c93b7397c7239ec99ec44b4923df31533b introduced a lot of
annoying error messages when configuring Qt. It makes qmake complain
that it cannot find rcc and uic, which aren't yet built at this point.

We're now checking for the existence of the depend_command if it is
given as absolute path. This is the case for uic and rcc.
Yeah this is a weak solution for sick code.

Task-number: QTBUG-13366
Reviewed-by: ossi