QS60style: itemview selection indication works incorrectly
Currently QS60Style shows tick mark for listWidget with multiselection
capability only after two or more items are selected. This is fixed
so that selection indication is shown in multiselection itemviews
(all of them and not just list widget) irregardless of number of
selected items.

Also, if itemview loses focus with singleitem itemview, style now
still draws item highlight to it (with 50% transparency).

Task-number: QTBUG-12875, QTBUG-13072
Reviewed-by: Janne Anttila