Fixes makeCurrent() failed messages on several systems.
Small but important fix for threaded environments. If a thread
exits without releasing the current context, it could cause
the resource system to get confused. Particularly on Windows, where
a DC can only be used in the thread it is retrieved in.
We don't need the EGL specific fix anymore since it will have the
same effect as a doneCurrent() call.