Make SourceOver work properly in Qt/DirectFB
Qt and DirectFB have different definitions of SourceOver. In DirectFB
it's defined as this:

/* pixel = (source * fs + destination * fd),
   sa = source alpha,
   da = destination alpha
*/

     DSPD_SRC_OVER       =  3, /* fs: 1.0     fd: 1.0-sa */

In other words, the source pixels are not multiplied by the alpha and
this leads to issues with RGBA values like 255, 0, 0, 0.

The composition mode that in DirectFB that maps onto Qt's
CompositionMode_SourceOver is DSPD_NONE, defined like this:

     DSPD_NONE           =  0, /* fs: sa      fd: 1.0-sa (defaults) */

Merge-request: 807
Reviewed-by: Donald Carr <donald.carr@nokia.com>