Panic in qs60style.cpp while drawing a QTreeView
QS60Style crashes when running a test application that modifies
active style AND application's palette. After switching to
other style, setting palette to application default, re-enabling the
style's palette and activating QS60Style, causes a crash. This is
due to that internal theme palette is tried to access, before it is
constructed.

Fixed by making sure that when theme palette is accessed, it is
not NULL.

Task-number: QTBUG-13553
Reviewed-by: Miikka Heikkinen