Make QStaticText honor QPainter::pen() for rich text on X11 w/raster
The document layou will call setPen() with the PaintContext's text
color. If this is different from the current painter settings, the pen
will be passed on to the paint engine. The result was that, on X11 using
the raster (or opengl) engine, QStaticText would record the pen as
changed even when it hadn't explicitly been changed by the document.

Task-number: QTBUG-13740
Reviewed-by: Samuel