Implement most of a push and pop scope functionality.
This patch fix some issues with scoping in QML (push and pop scope are
a private API, so no one else should use it). I'm not 100% sure of the
implementation as documentation is not so clear and autotests depends
on scopeChain function which hasn't been implemented yet. I believe
that it should work as "with" statement. I wrote few tests, but then
I found some incompatibility between this implementation and JSC based
one. I would say that it is good enough to go in, but we should keep
in mind that there are some refactoring needed in future as:
- use m_v8Scopes and m_v8Contexts as one set
- fix problems when setGlobalObject is called after pushScope (as far
I know it doesn't happen in QML but still...)