Partial revert of f599f2f9f7ff41fdcfda93b54abf7aef5b5c9b68
The comment was wrong, v8 value for undefined is kept by handler. I
couldn't reproduce situation in which function returns an empty handler
(I think it is not possible). I removed logic that change internal state
to invalid if given handler is empty. It was redundant and it presented
a cost on QSVP creation time. For safety I placed back an assert.