Mitigate private header problems in QtCreator by adding semi-private API
Add a semi-private API to get QScriptEngine for a QDeclarativeEngine. So far the qmljsdebugger lib in QtCreator
accessed the script engine via QDeclarativeEnginePrivate. Replace this by a minimal API that is still in a
private header, where we nevertheless can make some BC checks/guarantees.

Aaron Kennedy agreed with the idea.

Task-number: QTCREATORBUG-2179