QTBUG-13301: The menubar does now work when no window is showing
The reason is that actions from the menu bar have no target, and
is therefore sendt down the responder chain. This is sometimes
neccessary to make it work with modal dialogs. But this also
means that there will be no first responder window to catch the
action if no window is showing on screen. Luckily, in 4.8, we
added our own QNSApplication. And this will also be the last
resort in the responder chaing. Hence, we add code to handle
the menu bar action here.

Rev-by: carlos