Update the ICC mkspec: keep the stack aligned to 16-byte
After discussing with Intel, turns out that ICC defaults to aligning
the stack when it needs to (e.g., when issuing aligned operations), but
doesn't care otherwise. GCC, on the other hand, expects the stack to
always be aligned and will issue instructions without checking.

We'll probably add __attribute__((force_align_arg_pointer)) to some
functions in our code, but we won't be able to catch everything.

Reviewed-By: Bradley T. Hughes