Make the de-inlined isRightToLeft not get called from updateProperties
Before Qt 4.7, QString::isRightToLeft was an inline function that
called QString::updateProperties(). In Qt 4.7, QString::isRightToLeft
was de-inlined and is now called from QString::updateProperties().
According to the Binary Compatibility Guidelines, it's ok to
de-inline a function provided that it's ok the old method is called.

Under some rare circumstances nowadays, the old method could be called
from updateProperties(), which would result in an infinite loop
(updateProperties -> isRightToLeft -> updateProperties -> ...)

This is usually prevented by -fvisibility-inlines-hidden in GCC
(automatic in Qt) and also by -Wl,-Bsymbolic-functions (not automatic,
must pass -reduced-relocations to configure).

Reviewed-by: Bradley T. Hughes