Add support for ARMv7 atomic operations
ARMv7 adds the DMB (data memory barrier) instruction which we can use to
enforce memory barriers in QAtomicInt and QAtomicPointer. Other than
that, ARMv7 is identical to ARMv6.

Adjust the ARMv6 code to relax the compiler memory barriers on the
*Relaxed() operations, and use *Relaxed() functions together the
appropriate compiler barriers in the *Acquire(), *Release(), and
*Ordered() functions. For "pure" ARMv6 code, the barriers are only
compiler barriers, but for ARMv7, we also emit the DMB instruction.

Reviewed-by: thiago