Avoid OpenVG rendering errors when stroking an aliased path.
Stroking a path can sometimes result in inconsistent rendering
especially when combined with a clip. For example, if the logical edge
of a clip rect coincides with the logical edge of a path then it can
happen that the edge is not painted correctly because the stroke lies
outside the bounds of the clip rect. To workaround this problem, we add
the 'aliasedCoordinateDelta' such that the rounding will err on the
side of caution. This improves the correctness when using the raster
engine as a reference.

Task-number: QTBUG-13165
Reviewed-by: Samuel
(cherry picked from commit 74f09c794d5fdcb8f4b547b1c55bcd512ef07b95)