Setting QGraphicsItem::ItemClipsChildrenToShape forces ItemClipsToShape
Problem was that setting ItemClipsChildrenToShape on an item resulted
in item itself being clipped to its own shape (ItemClipsToShape).
This commit also reduces state changes on the painter and re-uses the
clip whenever possible, which in turn means better performance for
items that clip children to shape.

Auto test included.

Task-number: QTBUG-12760