Partial implementation of the QScriptClass.
This is wrong implementation of the class. We decided to push it only
because we need the functionality that it provides and we do not have
anything better right now.

Implementation approach is wrong because we won't be able to implement:
- QSV::setScriptClass
- QSV::scriptClass
- QScriptClassPropertyIterator

Implementation doesn't pass all autotests, have many FIXME and I don't
like it.

Reviewed-by: Olivier Goffart