Make QmlDebug protocol more robust
The protocol so far was client->server only. That is, there was no
sane way for a client to check whether a plugin on the server (service)
was available or not. E.g. calling Client::setEnabled(true) 'succeeded',
without a check whether there is actually a service to talk to.

The new protocol replaces this shortcoming by a service discovery
mechanism: Both client & service announce their available plugins at
handshake time, and later on if there are changes. The status is
reflected in Client::status() and Service::Status() , which are either

  NotConnected - no network connection, or not registered properly
  Unavailable - TCP/IP connection works, but no plugin with the same
name on the other side
  Enabled - You can connect to plugin on other side

The status changes happen automatically (no setEnabled() anymore).

Furthermore a version ID was added to the handshake, so that we can
extend the protocol further in the future :)
(cherry picked from commit fd9771c29d401d88779ab7c5d7715c9ca41dd723)