Disallow patching capabilities of executables
All of the assigned capabilities of executables are likely to be
actually needed unlike those of dlls, so do not create a patched
sis for an application with non-self-signable capabilities.
Similarly block creation of a patched package if any executable has
a protected range SID, as installer will refuse to install such a
package anyway.

Task-number: QTBUG-13886
Reviewed-by: Janne Koskinen